Added test folder and moved LoginComponent.tsx to the components folder.
