+++ frontend/src/app/(auth)/login/page.tsx
@@ -1,6 +1,10 @@
+/*
+    Renders the login component on the server side.
+*/
+
 import { auth } from "@/lib/auth";
 import { redirect } from "next/navigation";
-import LoginComponent from "./LoginComponent";
+import LoginComponent from "../../../components/LoginComponent";
 
 export default async function LoginPage() {
     const session = await auth();
